+++ src/main/java/com/example/demo/repository/NewsRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface NewsRepo extends JpaRepository<News, Integer> {
+public interface NewsRepo extends JpaRepository<News, Long> {
 
 }
